requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile ./requirements.in
#
aiohttp==3.8.5
    # via tuspy
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
blinker==1.6.2
    # via flask
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via flask
flask==2.3.3
    # via -r ./requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
future==0.18.3
    # via tuspy
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   -r ./requirements.in
    #   flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
pyvimeo==1.1.0
    # via -r ./requirements.in
requests==2.31.0
    # via
    #   pyvimeo
    #   tuspy
six==1.16.0
    # via tuspy
tinydb==4.8.0
    # via tuspy
tuspy==1.0.1
    # via pyvimeo
urllib3==2.0.4
    # via requests
werkzeug==2.3.7
    # via flask
yarl==1.9.2
    # via aiohttp
zipp==3.16.2
    # via importlib-metadata
